Issue #8 Second attempt at displaying balance -e Original/build commit ID: b35a5f56021bb5e707c6dbe287ee1d0182f0b4ec

+++ game_state.c
@@ -38,5 +38,5 @@ typedef struct {
     UINT8 hp;
     
     // Money the user has.
-    unsigned long balance;
+    unsigned int balance;
 } game_state_t;
